moved field initializer reading from custom ANTLR parser to Compiler Tree API
This sounds like it ties the implementation to the Sun Java Compiler,
but this tie was already in place with the previous implementation. In fact,
this implementation only relies on public APIs, even if they are
vendor-specific, while the old one relied on a private implementation detail
of a vendor-specific class.