separate @Inject from @Property
Turns out that in the real project, injecting properties via @Inject @Property
is not going to work. Instead, @Inject will be used for dependencies and
properties, marked with @Property, will be handled manually. This changes
the annotation processor to handle the new situation, and even issue a warning
when the combination of @Inject @Property is encountered.