document qualifier and scope annotations
Dependencies injected via @Inject can have qualifier annotations and scope
annotations that affect the injection process. Qualifier annotation is
an annotation that is itself annotated with @Qualifier. Scope annotation is
an annotation that is itself annotated with @Scope. These annotations are now
visible in the generated documentation, together with parameter values.