move build to Maven and add com.google.guava and javax.inject
The move to Maven was long overdue. With that, the classes from "framework"
that simulate the javax.inject and Guava classes were removed and the actual
classes are used instead.

The "annotation-processor" Maven module has to have annotation processing
turned off, otherwise the build fails. Fortunately, this is not a problem
in this case.