better index of units and services
The index of units and services now contains, for each class, a first line
of the Javadoc, stripped of all HTML tags. This requires an actual HTML parser,
doing things like this with a regexp is a direct road to hell. Hence
the dependency on jsoup, a nice HTML5 parser with sanitizing capabilities.