HTML output
This finally provides an almost-production-quality HTML output. The Markdown
output is gone because Markdown (at least the core Markdown) can't express
all the HTML tags that are needed (e.g. tables). Instead, HTML output
is generated directly from Mustache templates, using the excellent
Mustache.java library.

Bootstrap is used for styling, but a few very minor corrections had to be
applied (see style.css). It looks boring, but way better than unstyled HTML
and it's pretty usable. The static assets (currently only stylesheets, but
that may expand) are all copied to the generated output. They are small,
so it's not a big problem.