document all encountered classes and link to documentation from dependencies
An @Inject-ed dependency doesn't necessarily have to have any @Inject or
@Property annotation itself. Currently, such class wouldn't be documented
at all. This change adds a third pass of the documentation process that goes
through all the encountered classes and collects a documentation for those
that weren't documented in the previous pass.

Also, the documentation of a class now clearly says if each of the dependency
classes is documented. If it is, the "Documented" marker also works as a link.