a rudimentary renderer of Javadoc tags
The generated documentation contains Javadoc. That is generally fine,
as Javadoc is HTML, but there are special Javadoc tags like {@code ...}.
It would be nice if these were rendered to HTML, but that isn't exactly easy.
There is no standalone Javadoc parser and the one inside JDK is not reusable
directly. So for now, this only supports {@code ...} and {@literal ...},
because these can be easily rendered using regular expressions.