rewrite DocumentationGenerator to better conform to the lifecycle
When I first wrote DocumentationGenerator, I didn't fully understand
the annotation processor lifecycle, so I made a few simplifying assumptions
that turned out not to be true. Even if it works correctly in a simple
environment (no other annotation processor present), it is worth fixing,
epsecially since the fix is pretty small and localized.