add a Maven plugin
The AnnDocuGen annotation processor generates documentation, but the generated
documentation is just a simple directory and Maven can't work with it. That
is a problem in the real-world usecase. Ideally, it should work similarly to
how javadoc works in Maven. Luckily, that's pretty easy to achieve.

This adds a simple Maven plugin that creates a JAR archive of the generated
documentation and attaches it to the Maven project as an artifact. This means
that the generated documentation will be deployed to a Maven repository
automatically. The artifact has a classifier of "anndocugen".

To show how the plugin is used, it is applied to the example module.