move the actual process of writing files to a separate class
This is a first part of a bigger refactoring. It introduces a "view" package
that contains classes specific to actually generating documentation files.
This process shouldn't be dependent on the annotation processor environment,
which is why the process of creating output files is abstracted over using
the FileCreator interface.

The documentation collecting process and file writing process share the same
model classes for now. All collected data that should be written are stored
in the DocumentationData object.