introduce view-model classes
This is a third and final part of a bigger refactoring. It introduces
a view-model class for the DocumentedDependency model class and moves
the information whether the dependency class is documented from the model
class to the view-model class. This simplifies a lot of the logic, mainly
in that a lot of mutations are now avoided.

In addition to that, other view-model classes are introduced on the way from
the DocumentedDependency class to the root of the (view-)model hierarchy. All
the view-model classes are named after the respective model classes, appending
a "View" suffix. The view-model classes are used to generate the documentation
files instead of the model classes, but for the most part, the view-model
classes delegate to the model classes to avoid duplicity. Also, the view-model
classes mirror the interface of the model classes, so in fact, no changes to
the templates were needed.