add another module with examples
This adds a new Maven module "example2" with another set of examples. It will
be used to show the behavior of AnnDocuGen in a set of new situations that
were not envisioned previously (private or abstract classes, inheritance).
In addition to that, it will also be used to show how to generate unified
documentation of more Maven modules/projects (documentation merging).