document inherited properties and dependencies
For usability, it is important that the documentation shows inherited
properties and dependencies. This is solved on the view-model layer, which
required adding a view-model wrapper over DocumentedProperty (a wrapper for
DocumentedDependency was already in place). The only change in the model layer
was an addition of a parent type to DocumentedClass.

In the generated output, inherited properties and dependencies are marked with
a label "Inherited". That could possibly be a link to a documentation of
the class from which the property/dependency is inherited, but these parent
classes don't necessarily have to be documented.