add a new goal to the Maven plugin for AnnDocuGen aggregation
The Maven plugin has a new goal "aggregate" that looks through all direct
dependencies (i.e., not transitive) and collects all the AnnDocuGen-provided
data (raw-data.json). Once all dependencies are processed, aggregate
documentation is generated.

An example project "examples-dist" is provided that depends on "example"
and "example2" projects. It produces a zip file with all the dependencies
and an aggregate AnnDocuGen documentation.