configurable includes and excludes in the aggregate goal of the Maven plugin
The "aggregate" goal of the Maven plugin now has configuration parameters
called "includes" and "excludes". They are lists of dependency patterns
(like in the Maven Javadoc plugin) for artifacts that are either always
included or always excluded. If an artifact is always included, it must
contain AnnDocuGen output, otherwise the execution fails with an error.
If an artifact is always excluded, it is ignored and even if it contains
AnnDocuGen output, it won't be added to the aggregate.